+++ src/react/src/components/DashboardMergeFacilities.jsx
@@ -94,6 +94,7 @@ function DashboardMergeFacilities({
                     error={targetError}
                     handleEnterKeyPress={fetchTargetFacilityOnEnterKeyPress}
                     title="Target facility for merge"
+                    highlightBackground
                 />
                 <DashboardFacilityCard
                     updateOSID={updateToMergeOSID}
@@ -105,6 +106,7 @@ function DashboardMergeFacilities({
                     error={toMergeError}
                     handleEnterKeyPress={fetchToMergeFacilityOnEnterKeyPress}
                     title="Facility to merge into target"
+                    highlightBackground
                 />
             </div>
         </>